refactor: userpage to reusable components
